# Sparse gradient update in the SGD optimizer