@zippeurfou 2018-03-08
@entelechie What happened with this PR? 
It seems that https://github.com/keras-team/keras/issues/4365 still exist no?